requirements.txt
django
djangorestframework
requests
python-dateutil
